Level should be uint32_t, not int32_t

+++ Solver/Gaussian.cpp
@@ -640,7 +640,7 @@ Gaussian::gaussian_ret Gaussian::handle_matrix_confl(PropBy& confl, const matrix
         #endif
 
         uint32_t maxsublevel_at = UINT_MAX;
-        for (uint32_t i = 0, size = cla.size(); i != size; i++) if (solver.level[cla[i].var()] == (int32_t)curr_dec_level) {
+        for (uint32_t i = 0, size = cla.size(); i != size; i++) if (solver.level[cla[i].var()] == curr_dec_level) {
             uint32_t tmp = find_sublevel(cla[i].var());
             if (tmp >= maxsublevel) {
                 maxsublevel = tmp;
@@ -664,7 +664,7 @@ Gaussian::gaussian_ret Gaussian::handle_matrix_confl(PropBy& confl, const matrix
 
 Gaussian::gaussian_ret Gaussian::handle_matrix_prop_and_confl(matrixset& m, uint32_t last_row, PropBy& confl)
 {
-    int32_t maxlevel = std::numeric_limits<int32_t>::max();
+    uint32_t maxlevel = std::numeric_limits<uint32_t>::max();
     uint32_t size = UINT_MAX;
     uint32_t best_row = UINT_MAX;
 
@@ -676,7 +676,7 @@ Gaussian::gaussian_ret Gaussian::handle_matrix_prop_and_confl(matrixset& m, uint
             analyse_confl(m, row, maxlevel, size, best_row);
     }
 
-    if (maxlevel != std::numeric_limits<int32_t>::max())
+    if (maxlevel != std::numeric_limits<uint32_t>::max())
         return handle_matrix_confl(confl, m, size, maxlevel, best_row);
 
     #ifdef DEBUG_GAUSS
@@ -749,7 +749,7 @@ void Gaussian::cancel_until_sublevel(const uint32_t until_sublevel)
     #endif
 }
 
-void Gaussian::analyse_confl(const matrixset& m, const uint32_t row, int32_t& maxlevel, uint32_t& size, uint32_t& best_row) const
+void Gaussian::analyse_confl(const matrixset& m, const uint32_t row, uint32_t& maxlevel, uint32_t& size, uint32_t& best_row) const
 {
     assert(row < m.num_rows);
 
@@ -767,7 +767,7 @@ void Gaussian::analyse_confl(const matrixset& m, const uint32_t row, int32_t& ma
     #endif
     #endif
 
-    int32_t this_maxlevel = 0;
+    uint32_t this_maxlevel = 0;
     unsigned long int var = 0;
     uint32_t this_size = 0;
     while (true) {
@@ -789,7 +789,7 @@ void Gaussian::analyse_confl(const matrixset& m, const uint32_t row, int32_t& ma
                 || (this_maxlevel == maxlevel && this_size < size)
                 || (this_size <= 1)
             )) {
-        assert(maxlevel != std::numeric_limits<int32_t>::max());
+        assert(maxlevel != std::numeric_limits<uint32_t>::max());
 
         #ifdef VERBOSE_DEBUG
         cout << "(" << matrix_no << ")Other found conflict just as good or better.";
